Only install needed requirements
We currently install SBCL on all systems - but only this repository
needs it. Also, this repo needs far less packages than we install by
default and the version of SBCL that is installed is so out of date that
the tests no longer pass.

Add an bindep.txt file containing a cross-platform list of
dependencies needed for running included tests.

Convert SBCL version to one installed by CIM so that we always get an
up to date version.

This change is self-testing.

Related Change: If8b4e412f8ae48c6c0133634df4a95d10a876be3
Change-Id: Ie1ecb184bd825aea4dbb315cf125b991ae22d682